Tri dans le bootstrap-year-calendar.js, correction du fichier index.d.ts de bootstrap-year-calendar pour exporter un module
